move tests specific to tslibs.parsing
